# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BizPortal Multi-Division API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      summary: List expenses
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: categoryId, in: query, schema: { type: integer } }
        - { name: expenseType, in: query, schema: { type: string } }
        - { name: salesDocId, in: query, schema: { type: integer } }
        - { name: shipmentId, in: query, schema: { type: integer } }
        - { name: search, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Expense" }
    post:
      operationId: createExpense
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateExpenseBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
...
  /expenses/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      operationId: getExpense
      summary: Get expense detail
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExpenseDetail" }
    put:
      operationId: updateExpense
      summary: Update expense
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateExpenseBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
    delete:
      operationId: deleteExpense
      summary: Delete expense
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    ExpenseCategory:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        expenseAccountId: { type: integer, nullable: true }
        payableAccountId: { type: integer, nullable: true }
        requiresAttachment: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string }
      required: [id, name, code, requiresAttachment, isActive, createdAt]
...
    Expense:
      type: object
      properties:
        id: { type: integer }
        expenseNumber: { type: string }
        date: { type: string }
        vendorEmployee: { type: string, nullable: true }
        expenseType: { type: string, enum: [vendor_bill, reimbursement, internal, routine, kasbon, talangan, fixed_asset] }
        salesDocId: { type: integer, nullable: true }
        shipmentId: { type: integer, nullable: true }
        categoryId: { type: integer, nullable: true }
        categoryName: { type: string, nullable: true }
        sourceAccountId: { type: integer, nullable: true }
        sourceAccountName: { type: string, nullable: true }
        vendorId: { type: integer, nullable: true }
        vendorName: { type: string, nullable: true }
        description: { type: string, nullable: true }
        qty: { type: number }
        unit: { type: string, nullable: true }
        unitPrice: { type: number }
        subtotal: { type: number }
        taxRateId: { type: integer, nullable: true }
        taxAmount: { type: number }
        total: { type: number }
        currency: { type: string }
        status: { type: string, enum: [draft, submitted, approved, pending_approval, posted, paid, rejected] }
        notes: { type: string, nullable: true }
        entryId: { type: integer, nullable: true }
        expenseAccountId: { type: integer, nullable: true }
        payableAccountId: { type: integer, nullable: true }
        rejectionReason: { type: string, nullable: true }
        createdById: { type: string, nullable: true }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, expenseNumber, date, expenseType, qty, unitPrice, subtotal, taxAmount, total, currency, status, createdAt, updatedAt]
    ExpenseAttachment:
      type: object
      properties:
        id: { type: integer }
        expenseId: { type: integer }
        objectPath: { type: string }
        fileName: { type: string }
        contentType: { type: string, nullable: true }
        createdAt: { type: string }
      required: [id, expenseId, objectPath, fileName, createdAt]
    ExpenseDetail:
      allOf:
        - $ref: "#/components/schemas/Expense"
        - type: object
          properties:
            attachments:
              type: array
              items: { $ref: "#/components/schemas/ExpenseAttachment" }
            category:
              $ref: "#/components/schemas/ExpenseCategory"
              nullable: true
          required: [attachments]
    CreateExpenseBody:
      type: object
      properties:
        date: { type: string }
        vendorEmployee: { type: string, nullable: true }
        vendorId: { type: integer, nullable: true }
        expenseType: { type: string, enum: [vendor_bill, reimbursement, internal, routine, kasbon, talangan, fixed_asset] }
        salesDocId: { type: integer, nullable: true }
        shipmentId: { type: integer, nullable: true }
        categoryId: { type: integer, nullable: true }
        description: { type: string, nullable: true }
        qty: { type: number }
        unit: { type: string, nullable: true }
        unitPrice: { type: number }
        taxRateId: { type: integer, nullable: true }
        currency: { type: string }
        notes: { type: string, nullable: true }
        expenseAccountId: { type: integer, nullable: true }
        payableAccountId: { type: integer, nullable: true }
        sourceAccountId: { type: integer, nullable: true }
      required: [date]
...